Refactor memory management and cleanup
- Updated the `kpoolmemmgr_HCBv2.cpp` to streamline memory allocation by directly assigning the physical base address from the allocator.
- Removed unnecessary size checks in `in_heap_alloc` to simplify allocation logic.
- Modified `mixed_bitmap.cpp` to enhance bitmap initialization and removed the second stage initialization function.
- Adjusted `out_surfaces.cpp` to accommodate changes in the allocator's return type and improved error handling.
- Deleted the obsolete `transparent_pages.cpp` file to clean up unused code.
- Fixed stack boundary checks in `kthread_interfaces.cpp` to ensure proper stack usage.
- Added debug output for TSC frequency calculation in `time.cpp` to aid in diagnostics.
- Removed redundant Intel CPUID functions from `intel_cpuid.cpp` to reduce code complexity.
- Updated memory map in `phy_instance_1.json` to correct the type of a memory region from BootServicesData to ConventionalMemory.

+++ src/include/memory/kpoolmemmgr.h
@@ -215,7 +215,7 @@ public:
         public: 
         struct alignas(16) data_meta{//每个被分配的都有元信息以及相应魔数
             //是分配在堆内，后面紧接着就是数据
-            uint16_t data_size;
+            uint32_t data_size;
             alloc_flags_t alloc_flags;
             uint64_t magic;
         };
